Use string_view for word extraction

+++ src/main.cpp
@@ -12,7 +12,7 @@ int main(int argc, char * argv[])
         }
 
         boost::asio::io_service io_service;
-        mailgram::Server server(io_service, std::atoi(argv[1]));
+        mailgram::Server server(io_service, static_cast<short int>(std::atoi(argv[1])));
         io_service.run();
     } catch (std::exception & e) {
         std::cerr << "Exception: " << e.what() << "\n";
